requirements.txt
certifi==2020.6.20
cycler==0.10.0
imageio==2.9.0
joblib==0.17.0
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.18.1
pandas==1.1.3
Pillow==8.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.21.3
scipy==1.2.1
six==1.15.0
torch>=1.4.0
torchvision>=0.5.0
tqdm==4.54.0